Move fishing activities to its own slice
